refactor test helpers tht run qunit
