+++ packages/host/tests/helpers/setup-qunit.js
@@ -1,0 +1,13 @@
+import * as TestWaiters from '@ember/test-waiters';
+
+import * as QUnit from 'qunit';
+import { setup } from 'qunit-dom';
+
+import { useTestWaiters } from '@cardstack/runtime-common';
+
+export function setupQUnit() {
+  QUnit.dump.maxDepth = 20;
+  useTestWaiters(TestWaiters);
+  setup(QUnit.assert);
+  QUnit.config.autostart = false;
+}
