--- packages/host/tests/live-test.html
@@ -1,74 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Realm Test Runner</title>
-    <meta name="description" content="" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-    {{content-for "head"}}
-
-    <link rel="stylesheet" href="{{rootURL}}assets/vendor.css" />
-    <link rel="stylesheet" href="{{rootURL}}assets/@cardstack/host.css" />
-    <link rel="stylesheet" href="{{rootURL}}assets/test-support.css" />
-
-    {{content-for "head-footer"}}
-  </head>
-  <body>
-    {{content-for "body"}}
-
-    <div id="qunit"></div>
-    <div
-      id="live-test-status"
-      style="position: fixed; top: 8px; left: 8px; z-index: 9999; background: #fff; padding: 4px 8px; border: 1px solid #ccc;"
-    >
-      Live-test page loaded.
-    </div>
-    <button
-      id="live-test-start"
-      type="button"
-      style="position: fixed; top: 8px; right: 8px; z-index: 9999; background: #0b5; color: #fff; border: none; padding: 6px 10px; cursor: pointer;"
-    >
-      Start QUnit
-    </button>
-    <div id="qunit-fixture">
-      <div id="ember-testing-container">
-        <div id="ember-testing"></div>
-      </div>
-    </div>
-
-    <div id="ember-basic-dropdown-wormhole"></div>
-    <div id="ember-testing-loader"></div>
-
-    <script type="module">
-      import * as ContentTag from '{{rootURL}}assets/content-tag/standalone.js';
-      globalThis.ContentTagGlobal = ContentTag;
-    </script>
-    <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/test-support.js"></script>
-    <script>
-      // Prevent any auto-start (e.g. test-helper/tests.js) from running before
-      // live-test.js applies its filter. Restored inside live-test.js.
-      console.info('[live-test] HTML loaded');
-      window.__liveTestOriginalQUnitStart = window.QUnit && window.QUnit.start;
-      if (window.QUnit) {
-        window.QUnit.start = function () {
-          window.__liveTestStartBlocked = true;
-        };
-        console.info('[live-test] QUnit detected in HTML');
-      } else {
-        console.warn('[live-test] QUnit missing in HTML');
-      }
-      const statusEl = document.getElementById('live-test-status');
-      if (statusEl) {
-        statusEl.textContent = window.QUnit
-          ? 'Live-test HTML loaded. QUnit present.'
-          : 'Live-test HTML loaded. QUnit missing.';
-      }
-    </script>
-    <script src="{{rootURL}}assets/@cardstack/host.js"></script>
-    <script src="{{rootURL}}assets/tests.js"></script>
-
-    {{content-for "body-footer"}}
-  </body>
-</html>
