+++ packages/host/tests/test-helper.js
@@ -2,25 +2,22 @@ import Application from '@cardstack/host/app';
 import config from '@cardstack/host/config/environment';
 import * as QUnit from 'qunit';
 import { setApplication } from '@ember/test-helpers';
-import { setup } from 'qunit-dom';
+import { start as startEmberQUnit } from 'ember-qunit';
 import setupOperatorModeParametersMatchAssertion from '@cardstack/host/tests/helpers/operator-mode-parameters-match';
 import start from 'ember-exam/test-support/start';
-import { useTestWaiters } from '@cardstack/runtime-common';
-import * as TestWaiters from '@ember/test-waiters';
 // eslint-disable-next-line ember/no-test-import-export
-import './live-test';
+import { loadRealmTests } from './live-test';
+// eslint-disable-next-line ember/no-test-import-export
+import { setupQUnit } from './helpers/setup-qunit';
 
-const url = new URL(window.location.href);
-const isLiveTest =
-  url.pathname.endsWith('/live-test.html') || url.searchParams.has('liveTest');
+const application = Application.create({
+  ...config.APP,
+  rootElement: '#ember-testing',
+});
 
-if (!isLiveTest) {
-  QUnit.dump.maxDepth = 20;
-
-  useTestWaiters(TestWaiters);
-  setApplication(Application.create(config.APP));
-
-  setup(QUnit.assert);
+function setupHostTests() {
+  setApplication(application);
+  setupQUnit();
   setupOperatorModeParametersMatchAssertion(QUnit.assert);
 
   const urlParams = new URLSearchParams(window.location.search);
@@ -32,4 +29,23 @@ if (!isLiveTest) {
   }
 
   start();
+}
+
+function setupLiveTests() {
+  setApplication(application);
+  setupQUnit();
+
+  loadRealmTests(application).catch((error) => {
+    console.error('Failed to load realm tests', error);
+    QUnit.start(); //restarting test due to failure
+  });
+}
+
+// Single check — prevents double QUnit initialization (one QUnit instance only).
+const isLiveTest = new URL(window.location.href).searchParams.has('liveTest');
+
+if (isLiveTest) {
+  setupLiveTests();
+} else {
+  setupHostTests();
 }
